+++ lib/datasafe/lib/request/list-data-safe-private-endpoints-request.ts
@@ -17,14 +17,14 @@ export interface ListDataSafePrivateEndpointsRequest extends common.BaseRequest 
   /**
    * A filter to return only resources that match the specified compartment OCID.
    */
-  "compartmentId"?: string;
+  "compartmentId": string;
   /**
    * A filter to return only resources that match the specified display name.
    *
    */
   "displayName"?: string;
   /**
-   * A filter to return only the private endpoints that match the specified VCN OCID.
+   * A filter to return only resources that match the specified VCN OCID.
    */
   "vcnId"?: string;
   /**
@@ -36,7 +36,7 @@ export interface ListDataSafePrivateEndpointsRequest extends common.BaseRequest 
    */
   "limit"?: number;
   /**
-   * The beginning page from which the results start retrieving.
+   * For list pagination. The page token representing the page at which to start retrieving results. It is usually retrieved from a previous \"List\" call. For details about how pagination works, see [List Pagination](https://docs.cloud.oracle.com/en-us/iaas/Content/API/Concepts/usingapi.htm#nine).
    */
   "page"?: string;
   /**