+++ lib/mysql/lib/response/create-configuration-response.ts
@@ -21,16 +21,18 @@ export interface CreateConfigurationResponse {
   "location": string;
   /**
    * For optimistic concurrency control. See `if-match`.
+   *
    */
   "etag": string;
   /**
-   * Unique Oracle-assigned identifier for the request. If you need to contact Oracle about
-   * a specific request, please provide the request ID.
+   * Unique Oracle-assigned identifier for the request. If you need to contact
+   * Oracle about a particular request, please provide the request ID.
    *
    */
   "opcRequestId": string;
   /**
-   * OCID of the WorkRequest associated with this operation.
+   * Unique Oracle-assigned identifier for the asynchronous request. You can use this to query status of the asynchronous operation.
+   *
    */
   "opcWorkRequestId": string;
   /**