+++ lib/mysql/lib/response/create-db-system-response.ts
@@ -17,6 +17,7 @@ import common = require("oci-common");
 export interface CreateDbSystemResponse {
   /**
    * For optimistic concurrency control. See `if-match`.
+   *
    */
   "etag": string;
   /**
@@ -24,13 +25,14 @@ export interface CreateDbSystemResponse {
    */
   "location": string;
   /**
-   * Unique Oracle-assigned identifier for the request. If you need to contact Oracle about
-   * a specific request, please provide the request ID.
+   * Unique Oracle-assigned identifier for the request. If you need to contact
+   * Oracle about a particular request, please provide the request ID.
    *
    */
   "opcRequestId": string;
   /**
-   * OCID of the WorkRequest associated with this operation.
+   * Unique Oracle-assigned identifier for the asynchronous request. You can use this to query status of the asynchronous operation.
+   *
    */
   "opcWorkRequestId": string;
   /**