+++ lib/mysql/lib/response/get-work-request-response.ts
@@ -16,16 +16,17 @@ import common = require("oci-common");
 
 export interface GetWorkRequestResponse {
   /**
-   * For optimistic concurrency control. See `If-None-Match`.
+   * For optimistic concurrency control. See `if-match`.
+   *
    */
   "etag": string;
   /**
-   * Retry the request after the specified number of seconds.
+   * A decimal number representing the number of seconds the client should wait before polling this endpoint again.
    */
   "retryAfter": number;
   /**
-   * Unique Oracle-assigned identifier for the request. If you need to contact Oracle about
-   * a specific request, please provide the request ID.
+   * Unique Oracle-assigned identifier for the request. If you need to contact
+   * Oracle about a particular request, please provide the request ID.
    *
    */
   "opcRequestId": string;