+++ lib/mysql/lib/response/list-backups-response.ts
@@ -16,13 +16,16 @@ import common = require("oci-common");
 
 export interface ListBackupsResponse {
   /**
-   * Unique Oracle-assigned identifier for the request. If you need to contact Oracle about
-   * a specific request, please provide the request ID.
+   * Unique Oracle-assigned identifier for the request. If you need to contact
+   * Oracle about a particular request, please provide the request ID.
    *
    */
   "opcRequestId": string;
   /**
-   * Opaque token representing the next page of results.
+   * For pagination of a list of items. When paging through a list, if this header appears in the response,
+   * then a partial list might have been returned. Include this value as the `page` parameter for the
+   * subsequent GET request to get the next batch of items.
+   *
    */
   "opcNextPage": string;
   /**